requirements.txt
google-adk>=1.0.0
google-cloud-firestore>=2.19.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
